requirements.txt
requests==2.32.5
pandas==2.3.3
numpy==2.4.3
python-dotenv==1.2.2
fastapi==0.115.0
uvicorn==0.30.0
tzdata==2024.1
